main.c and input.c

+++ include/gba/syscall.h
@@ -53,6 +53,6 @@ void SoundBiasReset(void);
 
 void SoundBiasSet(void);
 
-s32 GetBiosChecksum(void);
+u16 GetBiosChecksum(void);
 
 #endif // GUARD_GBA_SYSCALL_H
